Refactor @compiled/babel-plugin-strip-runtime tests
