+++ packages/babel-plugin-strip-runtime/src/__tests__/strip-runtime-source-code.test.ts
@@ -1,10 +1,6 @@
-import { transformSync as babelTransformSync } from '@babel/core';
-import type { BabelFileResult } from '@babel/core';
-import compiledBabelPlugin from '@compiled/babel-plugin';
-import { format } from 'prettier';
-
-import stripRuntimeBabelPlugin from '../index';
 import type { BabelFileMetadata } from '../types';
+
+import { transform, transformSync } from './transform';
 
 // Mock out FS to avoid writing to disk
 // We aren't processing the result anyway, so no need for specifying the response
@@ -15,67 +11,6 @@ const testStyleSheetPath =
 const regexToFindRequireStatements =
   /(require\('@compiled\/webpack-loader\/css-loader!@compiled\/webpack-loader\/css-loader\/compiled-css\.css\?style=.*;)/g;
 const testSSR = true;
-
-type TransformOptions = {
-  styleSheetPath?: string;
-  compiledRequireExclude?: boolean;
-  run: 'both' | 'bake' | 'extract';
-  runtime: 'automatic' | 'classic';
-  extractStylesToDirectory?: { source: string; dest: string };
-  babelJSXPragma?: string;
-  babelJSXImportSource?: string;
-};
-
-const transformSync = (code: string, opts: TransformOptions): BabelFileResult | null => {
-  const { styleSheetPath, compiledRequireExclude, run, runtime, extractStylesToDirectory } = opts;
-  const bake = run === 'both' || run === 'bake';
-  const extract = run === 'both' || run === 'extract';
-
-  return babelTransformSync(code, {
-    babelrc: false,
-    configFile: false,
-    filename: '/base/src/app.tsx',
-    generatorOpts: {
-      sourceFileName: '../src/app.tsx',
-    },
-    plugins: [
-      ...(bake
-        ? [[compiledBabelPlugin, { importReact: runtime === 'classic', optimizeCss: false }]]
-        : []),
-      ...(extract
-        ? [
-            [
-              stripRuntimeBabelPlugin,
-              { styleSheetPath, compiledRequireExclude, extractStylesToDirectory },
-            ],
-          ]
-        : []),
-    ],
-    presets: [
-      [
-        '@babel/preset-react',
-        {
-          runtime,
-          ...(opts.babelJSXPragma ? { pragma: opts.babelJSXPragma } : {}),
-          ...(opts.babelJSXImportSource ? { importSource: opts.babelJSXImportSource } : {}),
-        },
-      ],
-    ],
-  });
-};
-
-const transform = (c: string, opts: TransformOptions): string => {
-  const fileResult = transformSync(c, opts);
-
-  if (!fileResult || !fileResult.code) {
-    throw new Error(`Missing fileResult: ${fileResult}`);
-  }
-
-  return format(fileResult.code, {
-    parser: 'babel',
-    singleQuote: true,
-  });
-};
 
 // This test suite is designed to test source code, which is also known as first-party code
 describe('babel-plugin-strip-runtime using source code', () => {
@@ -249,418 +184,8 @@ describe('babel-plugin-strip-runtime using source code', () => {
       });
     });
 
-    describe('with jsx pragma', () => {
+    describe('with extractStylesToDirectory', () => {
       it('work with classic jsx pragma', () => {
-        const codeWithPragma = `
-          /** @jsx jsx */
-          import { css, jsx } from '@compiled/react';
-
-          const Component = () => (
-            <div css={{ fontSize: 12, color: 'blue' }}>
-              hello world 2
-            </div>
-          );
-
-          const Component2 = () => (
-            <div css={css({ fontSize: 12, color: 'pink' })}>
-              hello world 2
-            </div>
-          );
-        `;
-
-        const actual = transform(codeWithPragma, {
-          run: 'both',
-          runtime: 'classic',
-        });
-
-        expect(actual).toMatchInlineSnapshot(`
-          "/* app.tsx generated by @compiled/babel-plugin v0.0.0 */
-          import * as React from 'react';
-          import { ax, ix } from '@compiled/react/runtime';
-          const Component = () =>
-            /*#__PURE__*/ React.createElement(
-              'div',
-              {
-                className: ax(['_1wyb1fwx _syaz13q2']),
-              },
-              'hello world 2'
-            );
-          const Component2 = () =>
-            /*#__PURE__*/ React.createElement(
-              'div',
-              {
-                className: ax(['_1wyb1fwx _syaz32ev']),
-              },
-              'hello world 2'
-            );
-          "
-        `);
-      });
-
-      it('work with classic jsx pragma (custom import)', () => {
-        const codeWithPragma = `
-          /** @jsx myJsx */
-          import { css, jsx as myJsx } from '@compiled/react';
-
-          const Component = () => (
-            <div css={{ fontSize: 12, color: 'blue' }}>
-              hello world 2
-            </div>
-          );
-
-          const Component2 = () => (
-            <div css={css({ fontSize: 12, color: 'pink' })}>
-              hello world 2
-            </div>
-          );
-        `;
-
-        const actual = transform(codeWithPragma, {
-          run: 'both',
-          runtime: 'classic',
-        });
-
-        expect(actual).toMatchInlineSnapshot(`
-          "/* app.tsx generated by @compiled/babel-plugin v0.0.0 */
-          import * as React from 'react';
-          import { ax, ix } from '@compiled/react/runtime';
-          const Component = () =>
-            /*#__PURE__*/ React.createElement(
-              'div',
-              {
-                className: ax(['_1wyb1fwx _syaz13q2']),
-              },
-              'hello world 2'
-            );
-          const Component2 = () =>
-            /*#__PURE__*/ React.createElement(
-              'div',
-              {
-                className: ax(['_1wyb1fwx _syaz32ev']),
-              },
-              'hello world 2'
-            );
-          "
-        `);
-      });
-
-      it('throws if runtime is classic, and pragma is set in babel config', () => {
-        const codeWithPragma = `
-          /** @jsx jsx */
-          import { css, jsx } from '@compiled/react';
-
-          const Component = () => (
-            <div css={{ fontSize: 12, color: 'blue' }}>
-              hello world 2
-            </div>
-          );
-
-          const Component2 = () => (
-            <div css={css({ fontSize: 12, color: 'pink' })}>
-              hello world 2
-            </div>
-          );
-        `;
-
-        expect(() =>
-          transform(codeWithPragma, {
-            run: 'both',
-            runtime: 'classic',
-            babelJSXPragma: 'jsx',
-          })
-        ).toThrow();
-      });
-
-      it('throws if runtime is classic, and pragma is set in babel config (custom import)', () => {
-        const codeWithPragma = `
-          /** @jsx myJsx */
-          import { css, jsx as myJsx } from '@compiled/react';
-
-          const Component = () => (
-            <div css={{ fontSize: 12, color: 'blue' }}>
-              hello world 2
-            </div>
-          );
-
-          const Component2 = () => (
-            <div css={css({ fontSize: 12, color: 'pink' })}>
-              hello world 2
-            </div>
-          );
-        `;
-
-        expect(() =>
-          transform(codeWithPragma, {
-            run: 'both',
-            runtime: 'classic',
-            babelJSXPragma: 'myJsx',
-          })
-        ).toThrow();
-      });
-
-      it("doesn't do anything to emotion's classic jsx pragma", () => {
-        const codeWithPragma = `
-          /** @jsx jsx */
-          import { css, jsx } from '@emotion/react';
-
-          const Component = () => (
-            <div css={{ fontSize: 12, color: 'blue' }}>
-              hello world 2
-            </div>
-          );
-
-          const Component2 = () => (
-            <div css={css({ fontSize: 12, color: 'pink' })}>
-              hello world 2
-            </div>
-          );
-        `;
-
-        const actual = transform(codeWithPragma, {
-          run: 'both',
-          runtime: 'classic',
-        });
-
-        expect(actual).toMatchInlineSnapshot(`
-          "/** @jsx jsx */
-          import { css, jsx } from '@emotion/react';
-          const Component = () =>
-            jsx(
-              'div',
-              {
-                css: {
-                  fontSize: 12,
-                  color: 'blue',
-                },
-              },
-              'hello world 2'
-            );
-          const Component2 = () =>
-            jsx(
-              'div',
-              {
-                css: css({
-                  fontSize: 12,
-                  color: 'pink',
-                }),
-              },
-              'hello world 2'
-            );
-          "
-        `);
-      });
-
-      it("doesn't do anything to emotion's classic jsx pragma (custom import)", () => {
-        const codeWithPragma = `
-          /** @jsx myJsx */
-          import { css, jsx as myJsx } from '@emotion/react';
-
-          const Component = () => (
-            <div css={{ fontSize: 12, color: 'blue' }}>
-              hello world 2
-            </div>
-          );
-
-          const Component2 = () => (
-            <div css={css({ fontSize: 12, color: 'pink' })}>
-              hello world 2
-            </div>
-          );
-        `;
-
-        const actual = transform(codeWithPragma, {
-          run: 'both',
-          runtime: 'classic',
-        });
-
-        expect(actual).toMatchInlineSnapshot(`
-          "/** @jsx myJsx */
-          import { css, jsx as myJsx } from '@emotion/react';
-          const Component = () =>
-            myJsx(
-              'div',
-              {
-                css: {
-                  fontSize: 12,
-                  color: 'blue',
-                },
-              },
-              'hello world 2'
-            );
-          const Component2 = () =>
-            myJsx(
-              'div',
-              {
-                css: css({
-                  fontSize: 12,
-                  color: 'pink',
-                }),
-              },
-              'hello world 2'
-            );
-          "
-        `);
-      });
-
-      it("errors when emotion's classic jsx pragma is mixed with compiled", () => {
-        const codeWithPragma = `
-          /** @jsx jsx */
-          import { css } from '@compiled/react';
-          import { jsx } from '@emotion/react';
-
-          const Component = () => (
-            <div css={{ fontSize: 12, color: 'blue' }}>
-              hello world 2
-            </div>
-          );
-
-          const Component2 = () => (
-            <div css={css({ fontSize: 12, color: 'pink' })}>
-              hello world 2
-            </div>
-          );
-        `;
-
-        expect(() =>
-          transform(codeWithPragma, {
-            run: 'both',
-            runtime: 'classic',
-          })
-        ).toThrow();
-      });
-
-      it('work with automatic jsx pragma', () => {
-        const codeWithPragma = `
-          /** @jsxImportSource @compiled/react */
-          import { css } from '@compiled/react';
-
-          const Component = () => (
-            <div css={{ fontSize: 12, color: 'blue' }}>
-              hello world 2
-            </div>
-          );
-
-          const Component2 = () => (
-            <div css={css({ fontSize: 12, color: 'pink' })}>
-              hello world 2
-            </div>
-          );
-        `;
-
-        const actual = transform(codeWithPragma, {
-          run: 'both',
-          runtime: 'automatic',
-        });
-
-        expect(actual).toMatchInlineSnapshot(`
-          "/* app.tsx generated by @compiled/babel-plugin v0.0.0 */
-          import { ax, ix } from '@compiled/react/runtime';
-          import { jsxs as _jsxs } from 'react/jsx-runtime';
-          import { jsx as _jsx } from 'react/jsx-runtime';
-          const Component = () =>
-            /*#__PURE__*/ _jsx('div', {
-              className: ax(['_1wyb1fwx _syaz13q2']),
-              children: 'hello world 2',
-            });
-          const Component2 = () =>
-            /*#__PURE__*/ _jsx('div', {
-              className: ax(['_1wyb1fwx _syaz32ev']),
-              children: 'hello world 2',
-            });
-          "
-        `);
-      });
-
-      it("doesn't do anything to emotion's automatic jsx pragma", () => {
-        const codeWithPragma = `
-          /** @jsxImportSource @emotion/react */
-          import { css } from '@emotion/react';
-
-          const Component = () => (
-            <div css={{ fontSize: 12, color: 'blue' }}>
-              hello world 2
-            </div>
-          );
-
-          const Component2 = () => (
-            <div css={css({ fontSize: 12, color: 'pink' })}>
-              hello world 2
-            </div>
-          );
-        `;
-
-        const actual = transform(codeWithPragma, {
-          run: 'both',
-          runtime: 'automatic',
-        });
-
-        expect(actual).toMatchInlineSnapshot(`
-          "/** @jsxImportSource @emotion/react */
-          import { css } from '@emotion/react';
-          import { jsx as _jsx } from '@emotion/react/jsx-runtime';
-          const Component = () =>
-            _jsx('div', {
-              css: {
-                fontSize: 12,
-                color: 'blue',
-              },
-              children: 'hello world 2',
-            });
-          const Component2 = () =>
-            _jsx('div', {
-              css: css({
-                fontSize: 12,
-                color: 'pink',
-              }),
-              children: 'hello world 2',
-            });
-          "
-        `);
-      });
-
-      it('works fine if runtime is automatic and pragma is set in babel config', () => {
-        const codeWithPragma = `
-          import { css, jsx } from '@compiled/react';
-
-          const Component = () => (
-            <div css={{ fontSize: 12, color: 'blue' }}>
-              hello world 2
-            </div>
-          );
-
-          const Component2 = () => (
-            <div css={css({ fontSize: 12, color: 'pink' })}>
-              hello world 2
-            </div>
-          );
-        `;
-
-        const actual = transform(codeWithPragma, {
-          run: 'both',
-          runtime: 'automatic',
-          babelJSXImportSource: '@compiled/react',
-        });
-
-        expect(actual).toMatchInlineSnapshot(`
-          "/* app.tsx generated by @compiled/babel-plugin v0.0.0 */
-          import { ax, ix } from '@compiled/react/runtime';
-          import { jsxs as _jsxs } from '@compiled/react/jsx-runtime';
-          import { jsx as _jsx } from '@compiled/react/jsx-runtime';
-          const Component = () =>
-            _jsx('div', {
-              className: ax(['_1wyb1fwx _syaz13q2']),
-              children: 'hello world 2',
-            });
-          const Component2 = () =>
-            _jsx('div', {
-              className: ax(['_1wyb1fwx _syaz32ev']),
-              children: 'hello world 2',
-            });
-          "
-        `);
-      });
-
-      it('work with classic jsx pragma with extractStylesToDirectory', () => {
         const codeWithPragma = `
           /** @jsx myJsx */
           import { css, jsx as myJsx } from '@compiled/react';
@@ -709,7 +234,7 @@ describe('babel-plugin-strip-runtime using source code', () => {
         `);
       });
 
-      it('work with automatic jsx pragma with extractStylesToDirectory', () => {
+      it('work with automatic jsx pragma', () => {
         const codeWithPragma = `
           /** @jsxImportSource @compiled/react */
           import { css } from '@compiled/react';
